# Make digest parsing faster by manually parsing the string