--- a/server/remote_cache/digest/digest_test.go
+++ b/server/remote_cache/digest/digest_test.go
@@ -43,16 +43,14 @@ func TestParseDownloadResourceName(t *testing.T) {
 			resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/",
 			wantError:    status.InvalidArgumentError(""),
 		},
-		// TODO(iain): fix
-		// { // Invalid size
-		// 	resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/1234b",
-		// 	wantError:    status.InvalidArgumentError(""),
-		// },
-		// TODO(iain): fix
-		// { // Trailing slashes
-		// 	resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/1234////////",
-		// 	wantError:    status.InvalidArgumentError(""),
-		// },
+		{ // Invalid size
+			resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/1234b",
+			wantError:    status.InvalidArgumentError(""),
+		},
+		{ // Trailing slashes
+			resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/1234////////",
+			wantError:    status.InvalidArgumentError(""),
+		},
 		{ // Bad compression type
 			resourceName: "/compressed-blobs/unknownCompressionType/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/1234",
 			wantError:    status.InvalidArgumentError(""),
@@ -61,11 +59,10 @@ func TestParseDownloadResourceName(t *testing.T) {
 			resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49ac/1234",
 			wantDRN:      digest.NewCASResourceName(&repb.Digest{Hash: "072d9dd55aacaa829d7d1cc9ec8c4b5180ef49ac", SizeBytes: 1234}, "", repb.DigestFunction_SHA1),
 		},
-		// TODO(iain): fix
-		// { // MD5
-		// 	resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b51/1234",
-		// 	wantDRN:      digest.NewCASResourceName(&repb.Digest{Hash: "072d9dd55aacaa829d7d1cc9ec8c4b51", SizeBytes: 1234}, "", repb.DigestFunction_MD5),
-		// },
+		{ // MD5
+			resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b51/1234",
+			wantDRN:      digest.NewCASResourceName(&repb.Digest{Hash: "072d9dd55aacaa829d7d1cc9ec8c4b51", SizeBytes: 1234}, "", repb.DigestFunction_MD5),
+		},
 		{ // SHA256
 			resourceName: "/blobs/072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d/1234",
 			wantDRN:      digest.NewCASResourceName(&repb.Digest{Hash: "072d9dd55aacaa829d7d1cc9ec8c4b5180ef49acac4a3c2f3ca16a3db134982d", SizeBytes: 1234}, "", repb.DigestFunction_SHA256),