This does away with the regular expression parsing in digest.go in favor of splitting the provided resource name on `/` and validating the individual pieces. There is probably still lots of room for improvement here, but according to the benchmark in https://github.com/buildbuddy-io/buildbuddy/pull/9405, this is a bit over 10x faster:

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 7 5800X 8-Core Processor             
                       │ /home/iain/old.txt │         /home/iain/new.txt          │
                       │       sec/op       │   sec/op     vs base                │
ParseDownloadString-16         57.675µ ± 3%   4.821µ ± 1%  -91.64% (p=0.000 n=40)

                       │ /home/iain/old.txt │          /home/iain/new.txt          │
                       │        B/op        │     B/op      vs base                │
ParseDownloadString-16         3.835Ki ± 0%   2.641Ki ± 0%  -31.14% (p=0.000 n=40)

                       │ /home/iain/old.txt │     /home/iain/new.txt     │
                       │     allocs/op      │ allocs/op   vs base        │
ParseDownloadString-16           69.00 ± 0%   56.00 ± 0%  -18.84% (n=40)
```

Fixes: https://github.com/buildbuddy-io/buildbuddy-internal/issues/5015